Efficiency improvements, break out text subtypes check.
